+++ widgets/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod avatar;
+pub mod badge;
 pub mod buttons;
 pub mod checkbox;
 pub mod common_widget;
@@ -23,8 +24,8 @@ pub mod tabs;
 pub mod transform_box;
 pub mod prelude {
   pub use super::{
-    avatar::*, buttons::*, checkbox::*, common_widget::*, divider::*, grid_view::*, icon::*,
-    input::*, label::*, layout::*, list::*, menu::*, path::*, progress::*, radio::*, router::*,
-    scrollbar::*, select_region::*, slider::*, switch::*, tabs::*, transform_box::*,
+    avatar::*, badge::*, buttons::*, checkbox::*, common_widget::*, divider::*, grid_view::*,
+    icon::*, input::*, label::*, layout::*, list::*, menu::*, path::*, progress::*, radio::*,
+    router::*, scrollbar::*, select_region::*, slider::*, switch::*, tabs::*, transform_box::*,
   };
 }