+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Please only add new entries below the [Unreleased](#unreleased---releasedate) he
 
 - **themes**: Material theme now registers icons to `svg_registry`. (#806 @M-Adoo)
 
+- **widgets**: add Badge widget for showing notifications, counts, or status information on top of another widget.(#805 @wjian23)
+
 ### Breaking
 
 - **themes**: Removed `IconTheme`, `fill_svgs!`, `svgs` and `material_svgs`. Use `svg_registry` to manage and access icons instead. (#806 @M-Adoo)